pyproject.toml
[project]
name = "TestProject"
version = "0.1.0"
description = "Test Project"
authors = [
    {name = "みず", email = "mizu.copo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mizu-common-py",
    "yt-dlp>=2024.0.0",
    "click>=8.0.0",
    "python-decouple>=3.8",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-decouple-stubs>=0.1",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
    "types-requests>=2.32.4.20250913",
    "vulture>=2.14",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E", "W", "F",   # 基本
    "I",             # import
    "B",             # bugbear
    "SIM",           # simplify
    "C4",            # comprehension
    "ARG",           # unused args
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
python_version = "3.14"
files = ["src", "tests"]
mypy_path = "src:stubs"

[tool.taskipy.tasks]
test = "ruff check --fix src tests stubs && ruff format src tests stubs && mypy && pytest"

[tool.uv]
package = true

[tool.uv.sources]
mizu-common-py = { git = "https://github.com/mizucopo/mizu-common-py.git" }
